Add mutable raster type.
The raster borrows the destination buffer mutably.  Users should
create a new raster when they want to decode a new image, and drop it
once they are done.

We have chosen to take a u8 memory slice even for 16-bpp rasters
because we expect that separating the buffer into u8 and u16 at the
type level would lead to a lot of code duplication in the user code.
e.g. the user would have to choose between allocating a u8/u16 buffer,
and then all subsequent functions need u8/u16 versions.